Tests: Do not assume empty vectors alias
Currently, TS uses Base.mightalias to throw an error if the two inputs to three-
arg transcode alias.
Two problems here: Base.mightalias is internal, and the new memory PR to Julia
changes aliasing of zero-length arrays.
The temporary fix for now is to remove the test that tests that zero-length
arrays trips the Base.mightalias. Later, let's fix this properly by not relying
on Base internals in TS tests.