Add Random as test dependency
The TestExt extension is loaded when Random and Test are loaded, and is used in
TS's tests. However, we forgot to load Random in the test suite.
This accidentally worked before, because Random was in the sysimage, but this
will not continue to be the case in the future.

See the comments in PR 152