preserve GC root inside Memory
We could preserve the ptr field also, but seems unlikely to make a
difference in practice and adds some mildly annoying duplication of
storage bytes. In Julia v1.11, this field could be replaced in whole or
in part by a Memory{UInt8} object allocation, but currently
`jl_string_to_genericmemory` is not even exposed as an API in Base.